requirements.txt
absl-py==0.11.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-auth==1.27.1
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
joblib==1.0.1
Markdown==3.3.4
numpy==1.20.1
oauthlib==3.1.0
Pillow==8.1.2
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.8.0
torchvision==0.9.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
